+++ packages/core/src/actions/cancel.ts
@@ -1,7 +1,6 @@
 import isDevelopment from '#is-development';
 import {
   AnyActorScope,
-  AnyActor,
   AnyMachineSnapshot,
   EventObject,
   MachineContext,
@@ -89,8 +88,8 @@ export function cancel<
   sendId: ResolvableSendId<TContext, TExpressionEvent, TParams, TEvent>
 ): CancelAction<TContext, TExpressionEvent, TParams, TEvent> {
   function cancel(
-    args: ActionArgs<TContext, TExpressionEvent, TEvent>,
-    params: TParams
+    _args: ActionArgs<TContext, TExpressionEvent, TEvent>,
+    _params: TParams
   ) {
     if (isDevelopment) {
       throw new Error(`This isn't supposed to be called`);