+++ packages/xstate-graph/src/deduplicatePaths.ts
@@ -37,6 +37,7 @@ export const deduplicatePaths = <
   pathLoop: for (const pathWithEventSequence of allPathsWithEventSequence) {
     // Check each existing superpath to see if the path is a subpath of it
     superpathLoop: for (const superpathWithEventSequence of superpathsWithEventSequence) {
+      // eslint-disable-next-line @typescript-eslint/no-for-in-array
       for (const i in pathWithEventSequence.eventSequence) {
         // Check event sequence to determine if path is subpath, e.g.:
         //